third_party/blink/web_tests/fast/html/script-allowed-types-languages-expected.txt

This page tests the allowed values for the type and language attributes of the <script> tag. Below you will see the allowed values.

Type: unspecified, Language: unspecified
Type: "", Language: unspecified
Type: "text/javascript", Language: unspecified
Type: "text/ecmascript", Language: unspecified
Type: "text/x-javascript", Language: unspecified
Type: "text/x-ecmascript", Language: unspecified
Type: "application/javascript", Language: unspecified
Type: "application/ecmascript", Language: unspecified
Type: "application/x-javascript", Language: unspecified
Type: "application/x-ecmascript", Language: unspecified
Type: "text/javascript1.0", Language: unspecified
Type: "text/javascript1.1", Language: unspecified
Type: "text/javascript1.2", Language: unspecified
Type: "text/javascript1.3", Language: unspecified
Type: "text/javascript1.4", Language: unspecified
Type: "text/javascript1.5", Language: unspecified
Type: "text/jscript", Language: unspecified
Type: "text/livescript", Language: unspecified
Type: "  text/javascript", Language: unspecified
Type: "text/javascript  ", Language: unspecified
Type: "  text/javascript  ", Language: unspecified
Type: unspecified, Language: ""
Type: unspecified, Language: "javascript"
Type: unspecified, Language: "javascript1.0"
Type: unspecified, Language: "javascript1.1"
Type: unspecified, Language: "javascript1.2"
Type: unspecified, Language: "javascript1.3"
Type: unspecified, Language: "javascript1.4"
Type: unspecified, Language: "javascript1.5"
Type: unspecified, Language: "javascript1.6"
Type: unspecified, Language: "javascript1.7"
Type: unspecified, Language: "ecmascript"
Type: unspecified, Language: "livescript"
Type: unspecified, Language: "jscript"
Type: "", Language: ""
Type: "", Language: "javascript"
Type: "", Language: "javascript1.0"
Type: "", Language: "javascript1.1"
Type: "", Language: "javascript1.2"
Type: "", Language: "javascript1.3"
Type: "", Language: "javascript1.4"
Type: "", Language: "javascript1.5"
Type: "", Language: "javascript1.6"
Type: "", Language: "javascript1.7"
Type: "", Language: "javascript1.8"
Type: "", Language: "javascript1.9"
Type: "", Language: "ecmascript"
Type: "", Language: "livescript"
Type: "", Language: "jscript"
Type: "", Language: "ebayScript"
Type: "", Language: "    "
Type: "", Language: "  javascript"
Type: "", Language: "javascript  "
Type: "", Language: "  javascript  "
Type: "", Language: "abcdefg"
When a type is specified, the language attribute should be ignored.
Type: "text/javascript", Language: ""
Type: "text/javascript", Language: "javascript"
Type: "text/javascript", Language: "javascript1.0"
Type: "text/javascript", Language: "javascript1.1"
Type: "text/javascript", Language: "javascript1.2"
Type: "text/javascript", Language: "javascript1.3"
Type: "text/javascript", Language: "javascript1.4"
Type: "text/javascript", Language: "javascript1.5"
Type: "text/javascript", Language: "javascript1.6"
Type: "text/javascript", Language: "javascript1.7"
Type: "text/javascript", Language: "javascript1.8"
Type: "text/javascript", Language: "javascript1.9"
Type: "text/javascript", Language: "ecmascript"
Type: "text/javascript", Language: "livescript"
Type: "text/javascript", Language: "jscript"
Type: "text/javascript", Language: "ebayScript"
Type: "text/javascript", Language: "    "
Type: "text/javascript", Language: "  javascript"
Type: "text/javascript", Language: "javascript  "
Type: "text/javascript", Language: "  javascript  "
Type: "text/javascript", Language: "abcdefg"
